--- a/Package.swift
+++ b/Package.swift
@@ -15,7 +15,7 @@ let package = Package(
         .package(url: "https://github.com/yonaskolb/JSONUtilities.git", from: "3.3.0"),
         .package(url: "https://github.com/kylef/Spectre.git", from: "0.7.0"),
         .package(url: "https://github.com/onevcat/Rainbow.git", from: "2.1.0"),
-        .package(url: "https://github.com/xcodeswift/xcproj.git", .revision("f6d733fc7cc1e6f6bc67da05f6ffadbfe721a8f0"))
+        .package(url: "https://github.com/xcodeswift/xcproj.git", from: "1.4.0")
     ],
     targets: [
         .target(name: "XcodeGen", dependencies: [