--- a/Sources/XcodeGenKit/XCProjExtensions.swift
+++ b/Sources/XcodeGenKit/XCProjExtensions.swift
@@ -46,7 +46,7 @@ extension PBXVariantGroup: GroupChild {
 extension PBXProj {
 
     public func printGroups() -> String {
-        guard let project = projects.first, let mainGroup = groups.getReference(project.mainGroup) else {
+        guard let project = objects.projects.first?.value, let mainGroup = objects.groups.getReference(project.mainGroup) else {
             return ""
         }
         return printGroup(group: mainGroup)
@@ -55,11 +55,11 @@ extension PBXProj {
     public func printGroup(group: PBXGroup) -> String {
         var string = group.childName
         for reference in group.children {
-            if let group = groups.getReference(reference) {
+            if let group = objects.groups.getReference(reference) {
                 string += "\n 📁  " + printGroup(group: group).replacingOccurrences(of: "\n ", with: "\n    ")
-            } else if let fileReference = fileReferences.getReference(reference) {
+            } else if let fileReference = objects.fileReferences.getReference(reference) {
                 string += "\n 📄  " + fileReference.childName
-            } else if let variantGroup = variantGroups.getReference(reference) {
+            } else if let variantGroup = objects.variantGroups.getReference(reference) {
                 string += "\n 🌎  " + variantGroup.childName
             }
         }
@@ -67,6 +67,6 @@ extension PBXProj {
     }
 
     public func getGroupChild(reference: String) -> GroupChild? {
-        return groups.getReference(reference) ?? fileReferences.getReference(reference) ?? variantGroups.getReference(reference)
+        return objects.groups.getReference(reference) ?? objects.fileReferences.getReference(reference) ?? objects.variantGroups.getReference(reference)
     }
 }