… Tests/Fixtures/TestProject/Project.xcodeproj/project.pbxproj