Required changes in order to adopt: https://github.com/xcodeswift/xcproj/pull/136

# Performance Overview & Results

This PR is primarily motivated by improving the performance in generating a large project. The improvements in this PR are largely by converting many properties in PBXProj (of xcproj) to dictionaries vs arrays which provides constant time lookups for hot code paths vs linear searches.

Control: (before any sort of lookup improvements):
https://github.com/xcodeswift/xcproj/commit/49aab802779cfb767a3c0f818ea042e79dd8b0a3

Optimized:
https://github.com/yonaskolb/XcodeGen/pull/158/commits/f2c4ebabfd57f813da444e276850e7f23dd847c9

## Project Complexity

aggregateTargets		0 	
buildConfigurations		200 	
buildFiles		31305 	
configurationLists		100 	
containerItemProxies		378 	
copyFilesBuildPhases		2 	
fileElements		0 	
fileReferences		9489 	
frameworksBuildPhases		12 	
groups		675 	
headersBuildPhases		0 	
nativeTargets		99 	
projects		1 
referenceProxies		0 	
resourcesBuildPhases		17 	
shellScriptBuildPhases		8 	
sourcesBuildPhases		99 	
targetDependencies		378 	
variantGroups		0 	
versionGroups		0 	

## Hardware: 
- OS: macOS High Sierra
- MacBook Pro (Retina, 13-inch, Early 2015)
- Processor: 3.1 GHz Intel Core i7
- Memory: 16 GB 1867 MHz DDR3
- Laptop is plugged in for all tests

## Test Script
- Delete existing project, kill any lingering processes
- Generate the full project 5x
- Drop the top / bottom outliers and average the remaining generation times.


### Before Optimizations (in seconds)
1: 88.1857219934464s
2: 119.523904979229s
3: 98.0418360233307s
4: 93.2748199701309s
5: 94.7779580354691s
*Average*: **~95 seconds** `((98+93+94)/3)`

### After Optimizations (in seconds)
1: 46.6170199513435s
2: 74.6580080389977s
3: 33.5623190402985s
4: 37.5278360247612s
5: 29.6541140079498s
*Average*: **~38 seconds** `((33+37+46)/3)`

### Comparison
*Average Percentage Improvement*: **60%** `((1 - (38/95)) * 100)`