@yonaskolb 2017-11-17
@rahul-malik could you post in the description the results of a performance test highlighting the time to generate your large project before and after this PR, including the amount of targets and source files?