@yonaskolb 2017-11-20
Awesome, thanks @rahul-malik! Were the tests run using a debug or a release build?