@rahul-malik 2017-11-20
@yonaskolb - Good question - these were all done with a release build.