@yonaskolb 2017-11-20
I find it interesting that there's so much variance in the time, especially post optimization. 29s vs 74secs is over twice as long. 
The inputs and outputs should be the same. Was your computer maybe busy at some points there, because I would think the times should be fairly stable? Your culling of the bottom and top gives us good averages but it's still interesting.